Move publish-lockfile tests to a dedicated file.
